refactor(networking): move data_source networking port to services path
