+++ huaweicloud/services/acceptance/vpc/data_source_huaweicloud_networking_port_v2_test.go
@@ -1,18 +1,19 @@
-package huaweicloud
+package vpc
 
 import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 )
 
 func TestAccNetworkingV2PortDataSource_basic(t *testing.T) {
 	resourceName := "data.huaweicloud_networking_port.gw_port"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckNetworkingV2PortDestroy,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccNetworkingV2PortDataSource_basic(),
